+++ examples/carbon-for-ibm-products/example-gallery/src/gallery-config/index.js
@@ -26,6 +26,7 @@ import imageCreateTearsheet from './CreateTearsheet--thumbnail.png';
 import imageCreateTearsheetNarrow from './CreateTearsheetNarrow--thumbnail.png';
 import imageDataSpreadsheet from './DataSpreadsheet--thumbnail.png';
 import imageDatagrid from './Datagrid--thumbnail.png';
+import imageDecorator from './Decorator--thumbnail.png';
 import imageDelimitedList from './DelimitedList--thumbnail.png';
 import imageDescriptionList from './DescriptionList--thumbnail.png';
 import imageEditInPlace from './EditInPlace--thumbnail.png';
@@ -144,6 +145,11 @@ const config = [
     label: 'Datagrid',
     directory: 'Datagrid',
     thumbnail: `url(${imageDatagrid}`,
+  },
+  {
+    label: 'Decorator',
+    directory: 'Decorator',
+    thumbnail: `url(${imageDecorator}`,
   },
   {
     label: 'Delimited-List',