+++ packages/core/story-structure.js
@@ -53,21 +53,35 @@ const s = [
             ],
           },
           { n: 'DataSpreadsheet', s: ['c/DataSpreadsheet'] },
-          { n: 'TruncatedList', s: ['c/TruncatedList'] },
-          { n: 'DelimitedList', s: ['c/DelimitedList'] },
+          {
+            n: 'Decorators',
+            s: [
+              'c/Decorator',
+              'c/DecoratorLink',
+              'c/DecoratorSingleButton',
+              'c/DecoratorDualButton',
+            ],
+          },
+          { n: 'Delimited list', s: ['c/DelimitedList'] },
           { n: 'Description list', s: ['c/DescriptionList'] },
           { n: 'Modified tabs', s: ['c/ModifiedTabs'] },
+          { n: 'Nav', s: ['c/Nav'] },
           { n: 'Options tile', s: ['c/OptionsTile'] },
           { n: 'Page header', s: ['c/PageHeader'] },
+          {
+            n: 'Search bar',
+            s: ['c/SearchBar', 'c/SearchBarScopes'],
+          },
           { n: 'Side panel', s: ['c/SidePanel'] },
-          { n: 'String formatter', s: ['c/StringFormatter'] },
-          { n: 'Tag set', s: ['c/TagSet'] },
-          { n: 'Tearsheet', s: ['c/Tearsheet', 'c/TearsheetNarrow'] },
-          { n: 'User avatar', s: ['c/UserAvatar'] },
           {
             n: 'StatusIndicator',
             s: ['c/StatusIndicator', 'c/StatusIndicatorStep'],
           },
+          { n: 'String formatter', s: ['c/StringFormatter'] },
+          { n: 'Tag set', s: ['c/TagSet'] },
+          { n: 'Tearsheet', s: ['c/Tearsheet', 'c/TearsheetNarrow'] },
+          { n: 'Truncated list', s: ['c/TruncatedList'] },
+          { n: 'User avatar', s: ['c/UserAvatar'] },
         ],
       },
       {
@@ -75,6 +89,7 @@ const s = [
         s: [
           { n: 'Add Select', s: ['c/SingleAddSelect', 'c/MultiAddSelect'] },
           { n: 'About modal', s: ['c/AboutModal'] },
+          { n: 'Cascade', s: ['c/Cascade'] },
           {
             n: 'Create flows',
             s: [
@@ -85,6 +100,7 @@ const s = [
               'c/CreateSidePanel',
             ],
           },
+          { n: 'Drag and drop', s: ['c/DragAndDrop'] },
           {
             n: 'Edit and update',
             s: [
@@ -120,15 +136,10 @@ const s = [
           { n: 'Notifications', s: ['c/NotificationsPanel'] },
           { n: 'Remove', s: ['c/RemoveModal'] },
           { n: 'Saving', s: ['c/Saving'] },
-          {
-            n: 'Search bar',
-            s: ['c/SearchBar', 'c/SearchBarScopes'],
-          },
           { n: 'Status icons', s: ['c/StatusIcon'] },
           { n: 'Toolbars', s: ['c/Toolbar'] },
           { n: 'User profile images', s: ['c/UserProfileImage'] },
           { n: 'Web terminal', s: ['c/WebTerminal'] },
-          { n: 'Cascade', s: ['c/Cascade'] },
         ],
       },
       {
@@ -142,12 +153,13 @@ const s = [
           'c/CancelableTextEdit',
           'c/Carousel',
           'c/ComboButton',
+          'c/DecoratorBase',
           'c/ExampleComponent',
           'c/FilterSummary',
-          'c/TooltipTrigger',
-          'c/TearsheetShell',
           'c/SimpleHeader',
           'c/SteppedAnimatedMedia',
+          'c/TooltipTrigger',
+          'c/TearsheetShell',
         ],
       },
       {
@@ -172,6 +184,7 @@ const s = [
           },
           'c/Checklist',
           'c/Guidebanner',
+          'c/Inlinetip',
           {
             n: 'Interstitial Screen',
             s: [
@@ -180,8 +193,7 @@ const s = [
               'c/InterstitialScreenViewModule',
             ],
           },
-          'c/InlineTip',
-          'c/NonLinearReading',
+          { n: 'Non-linear reading', s: ['c/NonLinearReading'] },
         ],
       },
     ],