+++ packages/ibm-products/src/components/Carousel/Carousel.stories.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2023, 2023
+ * Copyright IBM Corp. 2023, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,17 +7,14 @@
 
 import React from 'react';
 
-import {
-  getStoryTitle,
-  prepareStory,
-} from '../../global/js/utils/story-helper';
+import { prepareStory } from '../../global/js/utils/story-helper';
 
 import { Carousel } from '.';
 import styles from './_storybook-styles.scss';
 import DocsPage from './Carousel.docs-page';
 
 export default {
-  title: getStoryTitle(Carousel.displayName),
+  title: 'IBM Products/Internal/Carousel',
   component: Carousel,
   tags: ['autodocs'],
   parameters: {