+++ packages/ibm-products/src/components/DataSpreadsheet/DataSpreadsheet.stories.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2022, 2022
+ * Copyright IBM Corp. 2022, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,10 +7,7 @@
 
 import React, { useMemo, useState } from 'react';
 
-import {
-  getStoryTitle,
-  prepareStory,
-} from '../../global/js/utils/story-helper';
+import { prepareStory } from '../../global/js/utils/story-helper';
 
 import { DataSpreadsheet } from '.';
 import { generateData } from './utils/generateData';
@@ -19,7 +16,7 @@ import { generateData } from './utils/generateData';
 import styles from './_storybook-styles.scss';
 
 export default {
-  title: getStoryTitle(DataSpreadsheet.displayName),
+  title: 'IBM Products/Components/Data spreadsheet/DataSpreadsheet',
   component: DataSpreadsheet,
   tags: ['autodocs'],
   argTypes: {