+++ packages/ibm-products/src/components/HTTPErrors/HTTPError403/HTTPError403.stories.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2021, 2021
+ * Copyright IBM Corp. 2021, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,14 +7,11 @@
 
 import React from 'react';
 import { HTTPError403 } from '.';
-import {
-  getStoryTitle,
-  prepareStory,
-} from '../../../global/js/utils/story-helper';
+import { prepareStory } from '../../../global/js/utils/story-helper';
 import { StoryDocsPage } from '../../../global/js/utils/StoryDocsPage';
 
 export default {
-  title: getStoryTitle(HTTPError403.displayName),
+  title: 'IBM Products/Patterns/HTTP errors/HTTPError403',
   component: HTTPError403,
   tags: ['autodocs'],
   parameters: {