+++ packages/ibm-products/src/components/StatusIndicator/StatusIndicatorStep.stories.js
@@ -7,10 +7,7 @@
 
 import React from 'react';
 
-import {
-  getStoryTitle,
-  prepareStory,
-} from '../../global/js/utils/story-helper';
+import { prepareStory } from '../../global/js/utils/story-helper';
 
 import { StatusIndicatorStep } from '.';
 import mdx from './StatusIndicatorStep.mdx';
@@ -18,7 +15,7 @@ import mdx from './StatusIndicatorStep.mdx';
 import styles from './_storybook-styles.scss';
 
 export default {
-  title: getStoryTitle(StatusIndicatorStep.displayName),
+  title: 'IBM Products/Components/Status indicator/StatusIndicatorStep',
   component: StatusIndicatorStep,
   tags: ['autodocs'],
   argTypes: {