+++ packages/ibm-products/src/components/Tearsheet/Tearsheet.stories.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2020, 2021
+ * Copyright IBM Corp. 2020, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -33,17 +33,15 @@ import {
 } from '../ActionSet/actions.js';
 
 import { getDeprecatedArgTypes } from '../../global/js/utils/props-helper';
-import {
-  getStoryTitle,
-  prepareStory,
-} from '../../global/js/utils/story-helper';
+import { prepareStory } from '../../global/js/utils/story-helper';
 
 import styles from './_storybook-styles.scss';
+import { TearsheetNarrow } from './TearsheetNarrow';
 
 // import mdx from './Tearsheet.mdx';
 
 export default {
-  title: getStoryTitle(Tearsheet.displayName),
+  title: 'IBM Products/Components/Tearsheet',
   component: Tearsheet,
   tags: ['autodocs'],
   parameters: { styles /* docs: { page: mdx } */, layout: 'fullscreen' },
@@ -236,7 +234,7 @@ const Template = ({ actions, slug, ...args }) => {
 };
 
 // eslint-disable-next-line react/prop-types
-const StackedTemplate = ({ actions, slug, ...args }) => {
+const StackedTemplate = ({ mixedSizes, actions, slug, ...args }) => {
   const [open1, setOpen1] = useState(false);
   const [open2, setOpen2] = useState(false);
   const [open3, setOpen3] = useState(false);
@@ -284,6 +282,8 @@ const StackedTemplate = ({ actions, slug, ...args }) => {
     return action;
   });
 
+  const VariableSizeTearsheet = mixedSizes ? TearsheetNarrow : Tearsheet;
+
   return (
     <>
       <style>{`.${pkg.prefix}--tearsheet { opacity: 0 }`};</style>
@@ -303,9 +303,11 @@ const StackedTemplate = ({ actions, slug, ...args }) => {
         <Button onClick={() => setOpen2(!open2)}>
           Toggle&nbsp;tearsheet&nbsp;2
         </Button>
-        <Button onClick={() => setOpen3(!open3)}>
-          Toggle&nbsp;tearsheet&nbsp;3
-        </Button>
+        {!mixedSizes && (
+          <Button onClick={() => setOpen3(!open3)}>
+            Toggle&nbsp;tearsheet&nbsp;3
+          </Button>
+        )}
       </ButtonSet>
       <div ref={ref}>
         <Tearsheet
@@ -338,7 +340,7 @@ const StackedTemplate = ({ actions, slug, ...args }) => {
             />
           </div>
         </Tearsheet>
-        <Tearsheet
+        <VariableSizeTearsheet
           {...args}
           actions={wiredActions2}
           headerActions={
@@ -367,24 +369,26 @@ const StackedTemplate = ({ actions, slug, ...args }) => {
               labelText="Enter an important value here"
             />
           </div>
-        </Tearsheet>
-        <Tearsheet
-          {...args}
-          actions={wiredActions3}
-          title="Tearsheet 3"
-          open={open3}
-          onClose={() => setOpen3(false)}
-          selectorPrimaryFocus="#stacked-input-3"
-          slug={slug && sampleSlug}
-        >
-          <div className="tearsheet-stories__dummy-content-block">
-            Main content 3
-            <TextInput
-              id="stacked-input-3"
-              labelText="Enter an important value here"
-            />
-          </div>
-        </Tearsheet>
+        </VariableSizeTearsheet>
+        {!mixedSizes && (
+          <Tearsheet
+            {...args}
+            actions={wiredActions3}
+            title="Tearsheet 3"
+            open={open3}
+            onClose={() => setOpen3(false)}
+            selectorPrimaryFocus="#stacked-input-3"
+            slug={slug && sampleSlug}
+          >
+            <div className="tearsheet-stories__dummy-content-block">
+              Main content 3
+              <TextInput
+                id="stacked-input-3"
+                labelText="Enter an important value here"
+              />
+            </div>
+          </Tearsheet>
+        )}
       </div>
     </>
   );
@@ -453,6 +457,19 @@ export const fullyLoaded = prepareStory(Template, {
 export const stacked = prepareStory(StackedTemplate, {
   storyName: 'Stacking tearsheets',
   args: {
+    closeIconDescription,
+    description,
+    height: 'lower',
+    influencer,
+    label,
+    actions: 7,
+  },
+});
+
+export const stackedMixedSizes = prepareStory(StackedTemplate, {
+  storyName: 'Stacking tearsheets, different sizes',
+  args: {
+    mixedSizes: true,
     closeIconDescription,
     description,
     height: 'lower',