+++ packages/ibm-products/src/components/TooltipTrigger/TooltipTrigger.stories.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2023, 2023
+ * Copyright IBM Corp. 2023, 2024
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -7,10 +7,7 @@
 import React from 'react';
 import { Monster } from '@carbon/react/icons';
 
-import {
-  getStoryTitle,
-  prepareStory,
-} from '../../global/js/utils/story-helper';
+import { prepareStory } from '../../global/js/utils/story-helper';
 
 import { TooltipTrigger } from '.';
 // import { pkg } from '../../settings';
@@ -19,7 +16,7 @@ import { TooltipTrigger } from '.';
 import styles from './_storybook-styles.scss';
 
 export default {
-  title: getStoryTitle(TooltipTrigger.displayName),
+  title: 'IBM Products/Internal/TooltipTrigger',
   component: TooltipTrigger,
   parameters: {
     styles,